+++ packages/terra-disclosure-manager/src/DisclosureManager.jsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import AppDelegate from 'terra-app-delegate';
+import DisclosureManagerDelegate from './DisclosureManagerDelegate';
+import DisclosureManagerContext from './DisclosureManagerContext';
+import withDisclosureManager from './withDisclosureManager';
 
 const availableDisclosureSizes = {
   TINY: 'tiny',
@@ -11,48 +13,22 @@ const availableDisclosureSizes = {
   FULLSCREEN: 'fullscreen',
 };
 
-const availableDisclosureHeights = {
-  240: 240,
-  420: 420,
-  600: 600,
-  690: 690,
-  780: 780,
-  870: 870,
-  960: 960,
-  1140: 1140,
-};
-
-const availableDisclosureWidths = {
-  320: 320,
-  480: 480,
-  640: 640,
-  800: 800,
-  960: 960,
-  1120: 1120,
-  1280: 1280,
-  1440: 1440,
-  1600: 1600,
-  1760: 1760,
-  1920: 1920,
-};
+const arrayReducer = (mappingObject, value) => Object.assign({ [`${value}`]: value }, mappingObject);
+const availableDisclosureHeights = [240, 420, 600, 690, 780, 870, 960, 1140].reduce(arrayReducer, {});
+const availableDisclosureWidths = [320, 480, 640, 800, 960, 1120, 1280, 1440, 1600, 1760, 1920].reduce(arrayReducer, {});
 
 const defaultDimensions = { height: availableDisclosureHeights['690'], width: availableDisclosureWidths['1120'] };
 const defaultSize = availableDisclosureSizes.SMALL;
 
 const isValidDimensions = dimensions => availableDisclosureHeights[dimensions.height] && availableDisclosureWidths[dimensions.width];
 
-const isValidSize = size => availableDisclosureSizes[size.toUpperCase()];
+const isValidSize = size => !!availableDisclosureSizes[size.toUpperCase()];
 
 export { availableDisclosureSizes, availableDisclosureHeights, availableDisclosureWidths };
 
 const propTypes = {
   /**
-   * An AppDelegate instance that will be integrated with the DisclosureManager instance. The DisclosureManager will defer to it if unsupported
-   * actions occur.
-   */
-  app: AppDelegate.propType,
-  /**
-   * The child components that will be provided with an AppDelegate 'app' prop used to interact with the DisclosureManager instance.
+   * The child components that will be provided the disclosure functionality.
    */
   children: PropTypes.node,
   /**
@@ -64,6 +40,16 @@ const propTypes = {
    * utilize its 'app' prop and forward the request instead of handling the request itself.
    */
   supportedDisclosureTypes: PropTypes.array,
+  /**
+   * A boolean indicating whether or not the DisclosureManager should handle all nested disclosure requests. When enabled, the DisclosureManager will handle all
+   * disclose requests coming from disclosured components, regardless of the preferred disclosure type.
+   */
+  trapNestedDisclosureRequests: PropTypes.bool,
+  /**
+   * A DisclosureManagerDelegate instance provided by a parent DisclosureManager. This prop is automatically provided by `withDisclosureManager` and should not
+   * be explicitly given to the component.
+   */
+  disclosureManager: DisclosureManagerDelegate.propType,
 };
 
 const defaultProps = {
@@ -78,6 +64,7 @@ class DisclosureManager extends React.Component {
     const newState = Object.assign({}, state);
     newState.disclosureComponentKeys = Object.assign([], newState.disclosureComponentKeys);
     newState.disclosureComponentData = Object.assign({}, newState.disclosureComponentData);
+    newState.disclosureComponents = Object.assign([], newState.disclosureComponents);
 
     return newState;
   }
@@ -109,24 +96,55 @@ class DisclosureManager extends React.Component {
     this.onDismissResolvers = {};
 
     this.state = {
+      childComponents: this.renderContentComponents(props.disclosureManager, props.children),
       disclosureIsOpen: false,
-      disclosureIsFocused: true,
+      disclosureIsFocused: false,
       disclosureIsMaximized: false,
       disclosureSize: undefined,
       disclosureDimensions: undefined,
       disclosureComponentKeys: [],
       disclosureComponentData: {},
+      disclosureComponents: [],
     };
+  }
+
+  componentDidUpdate(prevProps) {
+    const { children, disclosureManager } = this.props;
+
+    if (children !== prevProps.children || disclosureManager !== prevProps.disclosureManager) {
+      /**
+       * With the future deprecation of componentWillReceiveProps, there is no existing lifecycle method
+       * that allows for old and new prop comparisons before rendering. componentDidUpdate can be
+       * used to ensure that the state is sycned with the props when necessary; however, if the state does
+       * need to be updated, a second render will need to occur. I do not anticipate this happening often
+       * given how the component is typically consumed.
+       *
+       * Usage of setState in componentDidUpdate is flagged by eslint due to the potential of infinite renders.
+       * However, this implementation of setState should be safe.
+       */
+      // eslint-disable-next-line react/no-did-update-set-state
+      this.setState({
+        childComponents: this.renderContentComponents(disclosureManager, children),
+        disclosureIsOpen: false,
+        disclosureIsFocused: false,
+        disclosureIsMaximized: false,
+        disclosureSize: undefined,
+        disclosureDimensions: undefined,
+        disclosureComponentKeys: [],
+        disclosureComponentData: {},
+        disclosureComponents: [],
+      });
+    }
   }
 
   /**
    * Determines if the provided disclosure type is supported by the DisclosureManager.
-   * @return `true` if the type is supported or if there is no fallback `app` present. `false` is returned otherwise.
+   * @return `true` if the type is supported or if there is no fallback `disclosureManager` present. `false` is returned otherwise.
    */
   disclosureTypeIsSupported(type) {
-    const { app, supportedDisclosureTypes } = this.props;
+    const { disclosureManager, supportedDisclosureTypes } = this.props;
 
-    return supportedDisclosureTypes.indexOf(type) >= 0 || !app;
+    return supportedDisclosureTypes.indexOf(type) >= 0 || !disclosureManager;
   }
 
   openDisclosure(data) {
@@ -148,7 +166,7 @@ class DisclosureManager extends React.Component {
       size = defaultSize;
     }
 
-    this.setState({
+    const newState = {
       disclosureIsOpen: true,
       disclosureIsFocused: true,
       disclosureSize: size,
@@ -162,7 +180,9 @@ class DisclosureManager extends React.Component {
           component: data.content.component,
         },
       },
-    });
+    };
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState, undefined, this.props.trapNestedDisclosureRequests) }));
   }
 
   pushDisclosure(data) {
@@ -175,6 +195,7 @@ class DisclosureManager extends React.Component {
       props: data.content.props,
       component: data.content.component,
     };
+    newState.disclosureComponents = newState.disclosureComponents.concat(this.renderDisclosureComponents(this.props.disclosureManager, newState, [data.content.key], this.props.trapNestedDisclosureRequests));
 
     this.setState(newState);
   }
@@ -183,6 +204,7 @@ class DisclosureManager extends React.Component {
     const newState = DisclosureManager.cloneDisclosureState(this.state);
 
     newState.disclosureComponentData[newState.disclosureComponentKeys.pop()] = undefined;
+    newState.disclosureComponents.pop();
 
     this.setState(newState);
   }
@@ -196,31 +218,36 @@ class DisclosureManager extends React.Component {
       disclosureDimensions: undefined,
       disclosureComponentKeys: [],
       disclosureComponentData: {},
+      disclosureComponents: [],
     });
   }
 
   requestDisclosureFocus() {
-    this.setState({
-      disclosureIsFocused: true,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsFocused = true;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState, undefined, this.props.trapNestedDisclosureRequests) }));
   }
 
   releaseDisclosureFocus() {
-    this.setState({
-      disclosureIsFocused: false,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsFocused = false;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState, undefined, this.props.trapNestedDisclosureRequests) }));
   }
 
   maximizeDisclosure() {
-    this.setState({
-      disclosureIsMaximized: true,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsMaximized = true;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState, undefined, this.props.trapNestedDisclosureRequests) }));
   }
 
   minimizeDisclosure() {
-    this.setState({
-      disclosureIsMaximized: false,
-    });
+    const newState = DisclosureManager.cloneDisclosureState(this.state);
+    newState.disclosureIsMaximized = false;
+
+    this.setState(Object.assign(newState, { disclosureComponents: this.renderDisclosureComponents(this.props.disclosureManager, newState, undefined, this.props.trapNestedDisclosureRequests) }));
   }
 
   /**
@@ -299,15 +326,20 @@ class DisclosureManager extends React.Component {
     };
   }
 
-  renderContentComponents() {
-    const { children, app } = this.props;
-
-    const appDelegate = {};
+  /**
+   * Generates an Array of component instances based on the given children. These components can be stored in state to ensure subsequent
+   * renders are more efficient and do not unnecessarily generate new DisclosureManagerDelegate instances.
+   * @param {DisclosureManagerDelegate} disclosureManager A DisclosureManagerDelegate instance that will be used as a fallback in cases where unsupported
+   *                                                      disclosure types are requested.
+   * @param {Children} children The children prop value that will be wrapped in the generated provider.
+   */
+  renderContentComponents(disclosureManager, children) {
+    const delegate = {};
 
     /**
      * The disclose function provided will open the disclosure with the provided content.
      */
-    appDelegate.disclose = (data) => {
+    delegate.disclose = (data) => {
       if (this.disclosureTypeIsSupported(data.preferredType)) {
         return this.safelyCloseDisclosure()
           .then(() => {
@@ -331,32 +363,44 @@ class DisclosureManager extends React.Component {
             };
           });
       }
-      return app.disclose(data);
+      return disclosureManager.disclose(data);
     };
 
-    return React.Children.map(children, child => React.cloneElement(child, {
-      app: AppDelegate.clone(app, appDelegate),
-    }));
+    return (
+      <DisclosureManagerContext.Provider value={DisclosureManagerDelegate.create(delegate)}>
+        {children}
+      </DisclosureManagerContext.Provider>
+    );
   }
 
-  renderDisclosureComponents() {
-    const { app } = this.props;
+  /**
+   * Generates an Array of component instances based on the given disclosure state. These components can be stored in state to ensure subsequent
+   * renders are more efficient and do not unnecessarily generate new DisclosureManagerDelegate instances.
+   * @param {DisclosureManagerDelegate} disclosureManager A DisclosureManagerDelegate instance that will be used as a fallback in cases where unsupported
+   *                                                      disclosure types are requested.
+   * @param {Object} disclosureState An Object representing the state of the disclosure manager.
+   * @param {Array} componentKeysOverride An Array of component keys representing the components that are to be generated. If not provided, all component keys
+   *                                      found in the disclosureState will be used for generation.
+   * @param {Boolean} trapNestedDisclosureRequests A Boolean value indicating whether or not disclosures for disclosure components should be strict.
+   */
+  renderDisclosureComponents(disclosureManager, disclosureState, componentKeysOverride, trapNestedDisclosureRequests) {
     const {
       disclosureComponentKeys, disclosureComponentData, disclosureIsMaximized, disclosureIsFocused, disclosureSize,
-    } = this.state;
+    } = disclosureState;
 
-    return disclosureComponentKeys.map((componentKey, index) => {
+    return (componentKeysOverride || disclosureComponentKeys).map((componentKey) => {
       const componentData = disclosureComponentData[componentKey];
       const isFullscreen = disclosureSize === availableDisclosureSizes.FULLSCREEN;
       const popContent = this.generatePopFunction(componentData.key);
+      const componentIndex = disclosureComponentKeys.indexOf(componentKey);
 
-      const disclosureApp = {};
+      const delegate = {};
 
       /**
        * The disclose function provided will push content onto the disclosure stack.
        */
-      disclosureApp.disclose = (data) => {
-        if (this.disclosureTypeIsSupported(data.preferredType)) {
+      delegate.disclose = (data) => {
+        if (trapNestedDisclosureRequests || this.disclosureTypeIsSupported(data.preferredType)) {
           return Promise.resolve()
             .then(() => {
               this.pushDisclosure(data);
@@ -369,86 +413,80 @@ class DisclosureManager extends React.Component {
               };
             });
         }
-        return app.disclose(data);
+        return disclosureManager.disclose(data);
       };
 
       /**
        * Allows a component to remove itself from the disclosure stack. If the component is the only element in the disclosure stack,
        * the disclosure is closed.
        */
-      disclosureApp.dismiss = index > 0 ? popContent : this.safelyCloseDisclosure;
+      delegate.dismiss = componentIndex > 0 ? popContent : this.safelyCloseDisclosure;
 
       /**
        * Allows a component to close the entire disclosure stack.
        */
-      disclosureApp.closeDisclosure = this.safelyCloseDisclosure;
+      delegate.closeDisclosure = this.safelyCloseDisclosure;
 
       /**
        * Allows a component to remove itself from the disclosure stack. Functionally similar to `dismiss`, however `onBack` is
        * only provided to components in the stack that have a previous sibling.
        */
-      disclosureApp.goBack = index > 0 ? popContent : undefined;
+      delegate.goBack = componentIndex > 0 ? popContent : undefined;
 
       /**
        * Allows a component to request focus from the disclosure in the event that the disclosure mechanism in use utilizes a focus trap.
        */
-      disclosureApp.requestFocus = disclosureIsFocused ? () => Promise.resolve().then(this.releaseDisclosureFocus) : undefined;
+      delegate.requestFocus = disclosureIsFocused ? () => Promise.resolve().then(this.releaseDisclosureFocus) : undefined;
 
       /**
        * Allows a component to release focus from itself and return it to the disclosure.
        */
-      disclosureApp.releaseFocus = !disclosureIsFocused ? () => Promise.resolve().then(this.requestDisclosureFocus) : undefined;
+      delegate.releaseFocus = !disclosureIsFocused ? () => Promise.resolve().then(this.requestDisclosureFocus) : undefined;
 
       /**
        * Allows a component to maximize its presentation size. This is only provided if the component is not already maximized.
        */
-      disclosureApp.maximize = (!isFullscreen && !disclosureIsMaximized) ? () => (Promise.resolve().then(this.maximizeDisclosure)) : undefined;
+      delegate.maximize = (!isFullscreen && !disclosureIsMaximized) ? () => (Promise.resolve().then(this.maximizeDisclosure)) : undefined;
 
       /**
        * Allows a component to minimize its presentation size. This is only provided if the component is currently maximized.
        */
-      disclosureApp.minimize = (!isFullscreen && disclosureIsMaximized) ? () => (Promise.resolve().then(this.minimizeDisclosure)) : undefined;
+      delegate.minimize = (!isFullscreen && disclosureIsMaximized) ? () => (Promise.resolve().then(this.minimizeDisclosure)) : undefined;
 
       /**
        * Allows a component to register a function with the DisclosureManager that will be called before the component is dismissed for any reason.
        */
-      disclosureApp.registerDismissCheck = (checkFunc) => {
+      delegate.registerDismissCheck = (checkFunc) => {
         this.dismissChecks[componentData.key] = checkFunc;
 
-        if (app && app.registerDismissCheck) {
+        if (disclosureManager && disclosureManager.registerDismissCheck) {
           // The combination of all managed dismiss checks is registered to the parent app delegate to ensure
           // that all are accounted for by the parent.
-          return app.registerDismissCheck(() => Promise.all(Object.values(this.dismissChecks)));
+          return disclosureManager.registerDismissCheck(() => Promise.all(Object.values(this.dismissChecks)));
         }
 
         return Promise.resolve();
       };
 
-      if (componentData.component) {
-        return React.cloneElement(componentData.component, {
-          key: componentData.key,
-          app: AppDelegate.create(disclosureApp),
-        });
-      }
-
-      const ComponentClass = AppDelegate.getComponentForDisclosure(componentData.name);
-      if (!ComponentClass) {
-        return undefined;
-      }
-
-      return <ComponentClass key={componentData.key} {...componentData.props} app={AppDelegate.create(disclosureApp)} />;
+      return (
+        <DisclosureManagerContext.Provider value={DisclosureManagerDelegate.create(delegate)} key={componentData.key}>
+          {componentData.component}
+        </DisclosureManagerContext.Provider>
+      );
     });
   }
 
   render() {
     const { render } = this.props;
     const {
+      childComponents,
       disclosureIsOpen,
       disclosureIsFocused,
       disclosureIsMaximized,
       disclosureSize,
       disclosureDimensions,
       disclosureComponentKeys,
+      disclosureComponents,
     } = this.state;
 
     if (!render) {
@@ -459,7 +497,7 @@ class DisclosureManager extends React.Component {
       dismissPresentedComponent: (disclosureComponentKeys.length > 1) ? this.generatePopFunction(disclosureComponentKeys[disclosureComponentKeys.length - 1]) : this.safelyCloseDisclosure,
       closeDisclosure: this.safelyCloseDisclosure,
       children: {
-        components: this.renderContentComponents(),
+        components: childComponents,
       },
       disclosure: {
         isOpen: disclosureIsOpen,
@@ -467,7 +505,7 @@ class DisclosureManager extends React.Component {
         isMaximized: disclosureIsMaximized,
         size: disclosureSize,
         dimensions: disclosureDimensions,
-        components: this.renderDisclosureComponents(),
+        components: disclosureComponents,
       },
     });
   }
@@ -476,4 +514,7 @@ class DisclosureManager extends React.Component {
 DisclosureManager.propTypes = propTypes;
 DisclosureManager.defaultProps = defaultProps;
 
-export default DisclosureManager;
+const disclosureManagerShape = DisclosureManagerDelegate.propType;
+
+export default withDisclosureManager(DisclosureManager);
+export { withDisclosureManager, disclosureManagerShape };