+++ packages/terra-slide-panel-manager/src/terra-dev-site/test/slide-panel-manager/SlidePanelManagerDefault.test.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
-import TestExample from 'terra-disclosure-manager/lib/terra-dev-site/test/common/TestExample';
+import DisclosureComponent from './DisclosureComponent';
 import SlidePanelManager from '../../../SlidePanelManager';
 
 const SlidePanelManagerDefault = () => (
   <div role="main" style={{ height: '100%' }}>
     <SlidePanelManager>
-      <TestExample identifier="root-component" disclosureType="panel" />
+      <DisclosureComponent identifier="root-component" disclosureType="panel" />
     </SlidePanelManager>
   </div>
 );