+++ packages/terra-application-name/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Changed
 * Wdio tests changed to use themeCombinationOfCustomProperties
 
+### Removed
+* Removed dependency on terra-app-delegate package.
+
 2.20.0 - (November 15, 2018)
 ------------------
 ### Changed